# internal/db/repository/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.27.0

package repository

import (
	"time"

	"github.com/google/uuid"
	"github.com/jackc/pgx/v5/pgtype"
)

type Book struct {
	ID        uuid.UUID `json:"id"`
	Title     string    `binding:"required" json:"title"`
	Author    string    `binding:"required" json:"author"`
	CreatedAt time.Time `json:"createdAt"`
	UpdatedAt time.Time `json:"updatedAt"`
}

type Profile struct {
	ID         uuid.UUID   `json:"id"`
	UserID     pgtype.UUID `binding:"required,uuid" json:"userId"`
	ProfilePic *string     `json:"profilePic"`
	Name       *string     `json:"name"`
	CoverPic   *string     `json:"coverPic"`
	CreatedAt  time.Time   `json:"createdAt"`
	UpdatedAt  time.Time   `json:"updatedAt"`
}

type User struct {
	ID        uuid.UUID `json:"id"`
	Name      *string   `json:"name"`
	Username  *string   `json:"username"`
	Email     string    `binding:"required,email" json:"email"`
	Password  string    `binding:"required" json:"password"`
	CreatedAt time.Time `json:"createdAt"`
	UpdatedAt time.Time `json:"updatedAt"`
}